remove the dist task
